+++ lib/libexec/optparser.sh
@@ -7,7 +7,7 @@
 # shellcheck source=lib/getoptions_abbr.sh
 . "$SHELLSPEC_LIB/getoptions_abbr.sh"
 
-# shellcheck disable=SC1083
+# shellcheck disable=SC1083,SC2016
 parser_definition() {
   extension "$@"
   set -- "$1" "$2" "error_handler ${3:-echo}"
@@ -78,8 +78,12 @@ parser_definition() {
   param LOGFILE --log-file init:='/dev/tty' -- \
     'Log file for %logger directive and trace [default: /dev/tty]'
 
-  flag KEEP_TEMPDIR --keep-tempdir -- \
+  param TMPDIR --tmpdir init:="${TMPDIR:-${TMP:-/tmp}}" -- \
+    'Specify temporary directory [default: $TMPDIR, $TMP or /tmp]'
+
+  flag KEEP_TMPDIR --keep-tmpdir -- \
     'Do not cleanup temporary directory [default: disabled]'
+  flag KEEP_TMPDIR --keep-tempdir validate:'deprecated $1' abbr: hidden:true
 
   msg -- '' '  **** Execution ****' ''
 
@@ -409,4 +413,11 @@ error_handler() {
   esac
   "$1" "$2"
   return 1
+}
+
+deprecated() {
+  case $1 in
+    --keep-tempdir)
+      warn "--keep-tempdir is deprecated. replace with --keep-tmpdir."
+  esac
 }